package/requirements.txt
urllib
requests
pandas
schedule
tweepy
bs4
python-dotenv